% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_default_formatting.R
\name{gt_default_formatting}
\alias{gt_default_formatting}
\title{Create a gt object with your preferential defaults}
\usage{
gt_default_formatting(gt_object, key_col, metric_cols, pct_cols, ...)
}
\arguments{
\item{gt_object}{a gt object}

\item{metric_cols}{columns from gt object that should be formatted
as numbers. Defaults to zero decimals.}

\item{pct_cols}{columns from gt object that should be formatted as percents.
Defaults to zero decimals.}

\item{...}{other args to pass}
}
\value{
returns a gt object with columns formatted
}
\description{
Create a gt object with your preferential defaults
}
\examples{
library(gt)
library(gtTemplates)
library(dplyr)
gt_default_formatting(gt::gt(us_migration))
gt_default_formatting(gt::gt(us_migration), metric_cols = c("m_births", "m_deaths"))
gt_default_formatting(gt::gt(mtcars), pct_cols = "dd")
}
