% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_simple.R
\name{gt_theme_simple}
\alias{gt_theme_simple}
\title{A basic gt table theme that captures personal preferences}
\usage{
gt_theme_simple(gt_object, ...)
}
\arguments{
\item{gt_object}{a gt object created by the gt package}

\item{...}{additional arguments to pass}
}
\value{
A 'gt_tbl' object with the specified theme applied
}
\description{
A basic gt table theme that captures personal preferences
}
\examples{
mtcars |>
gt::gt() |>
gt_theme_simple()
}
